--- src/main/java/racinggame/core/gasstation/GasStation.java
@@ -1,6 +1,0 @@
-package racinggame.core.gasstation;
-
-public interface GasStation {
-
-    int sellFuel();
-}
