--- src/main/java/racinggame/core/gasstation/TankGasStation.java
@@ -1,20 +1,0 @@
-package racinggame.core.gasstation;
-
-import java.util.Queue;
-import racinggame.exception.EmptyTankException;
-
-public class TankGasStation implements GasStation {
-
-    private final Queue<Integer> tank;
-
-    public TankGasStation(Queue<Integer> tank) {
-        this.tank = tank;
-    }
-
-    public int sellFuel() {
-        if (tank.peek() == null) {
-            throw new EmptyTankException();
-        }
-        return tank.poll();
-    }
-}
