+++ src/main/java/racinggame/domain/exception/EmptyTankException.java
@@ -1,0 +1,8 @@
+package racinggame.domain.exception;
+
+public class EmptyTankException extends RuntimeException {
+
+    public EmptyTankException() {
+        super("연료가 떨어졌습니다.");
+    }
+}
