+++ src/main/java/racinggame/domain/exception/InvalidCarRegistrationException.java
@@ -1,0 +1,9 @@
+package racinggame.domain.exception;
+
+public class InvalidCarRegistrationException extends RuntimeException {
+
+    public InvalidCarRegistrationException() {
+        super("차량은 한 대 이상이어야 합니다.");
+    }
+
+}
