+++ src/main/java/racinggame/domain/exception/InvalidUnknownCarNameException.java
@@ -1,0 +1,8 @@
+package racinggame.domain.exception;
+
+public class InvalidUnknownCarNameException extends RuntimeException {
+
+    public InvalidUnknownCarNameException() {
+        super("등록되지 않은 차량입니다.");
+    }
+}
