+++ src/main/java/racinggame/domain/gasstation/GasStation.java
@@ -1,0 +1,6 @@
+package racinggame.domain.gasstation;
+
+public interface GasStation {
+
+    int sellFuel();
+}
