--- src/main/java/racinggame/exception/EmptyTankException.java
@@ -1,8 +1,0 @@
-package racinggame.exception;
-
-public class EmptyTankException extends RuntimeException {
-
-    public EmptyTankException() {
-        super("연료가 떨어졌습니다.");
-    }
-}
