--- src/main/java/racinggame/exception/IntegerMismatchException.java
@@ -1,8 +1,0 @@
-package racinggame.exception;
-
-public class IntegerMismatchException extends RuntimeException {
-
-    public IntegerMismatchException() {
-        super("정수 값을 입력하세요.");
-    }
-}
