--- src/main/java/racinggame/exception/InvalidUnknownCarNameException.java
@@ -1,8 +1,0 @@
-package racinggame.exception;
-
-public class InvalidUnknownCarNameException extends RuntimeException {
-
-    public InvalidUnknownCarNameException() {
-        super("등록되지 않은 차량입니다.");
-    }
-}
